Move StrataPixel3DBlock into layers.py
Per review: it's a building-block module, so it belongs alongside the other
Strata blocks in layers.py rather than in strata.py (which now holds only the
composed Strata model and its metadata). Pure relocation -- the class is
unchanged; strata.py imports it from .layers, and the now-unused F / Mlp /
DepthwiseConv / Natten3DSelfAttention imports there are dropped.
