+++ physicsnemo/experimental/models/strata/layers.py
@@ -33,7 +33,7 @@ from __future__ import annotations
 
 import warnings
 from functools import partial
-from typing import Optional, Tuple, Union
+from typing import Literal, Optional, Tuple, Union
 
 import torch
 import torch.nn as nn
@@ -48,6 +48,7 @@ from physicsnemo.nn.module.mlp_layers import Mlp
 __all__ = [
     "Natten3DSelfAttention",
     "StrataTransformer3DBlock",
+    "StrataPixel3DBlock",
     "PatchEmbed3D",
     "FinalLayer3D",
     "RopeTables",
@@ -685,3 +686,362 @@ class DepthwiseConv(torch.nn.Conv2d):
             )
 
         return super().forward(x, *args, **kwargs)
+
+
+class StrataPixel3DBlock(nn.Module):
+    r"""Pixel-pathway transformer block with pixel-wise adaptive layer norm.
+
+    Like a DiT block, but the adaptive-layer-norm (AdaLN) modulation parameters
+    (``shift``, ``scale``, ``gate`` for both the attention and MLP sub-layers)
+    vary *per pixel* and are derived from the backbone conditioning tokens
+    ``backbone_cond``. This is a **regression** conditioning: unlike the original DiT's
+    adaLN-Zero (which modulates on the diffusion timestep / noise level), here
+    the modulation is a learned function of the backbone stage's features —
+    there is no diffusion timestep or noise. Two derivation modes are supported
+    (``adaln_mode``):
+
+    - ``"pixel_proj"``: project each backbone token to
+      ``pixels_per_patch * 6 * dim`` values and scatter them to the pixels of
+      that patch.
+    - ``"bilinear_dw"``: trilinearly upsample the backbone tokens to pixel
+      resolution (over depth and the horizontal plane), smooth horizontal patch
+      seams with a per-level depthwise :math:`5\times5` convolution,
+      RMS-normalize, apply GeLU, then project to ``6 * dim``.
+
+    Modulation projections are zero-initialized so the pixel pathway starts as
+    an identity residual mapping.
+
+    The pixel-wise AdaLN is an independent reimplementation of the conditioning
+    in Strata (see :class:`Strata`); the ``"bilinear_dw"`` mode is an
+    original addition beyond that work.
+
+    Parameters
+    ----------
+    dim : int
+        Pixel-pathway embedding dimension.
+    cond_dim : int
+        Backbone-pathway (conditioning) embedding dimension.
+    pixels_per_patch : int
+        Number of pixels per backbone patch (:math:`p_d \cdot p_h \cdot p_w`);
+        used by the ``"pixel_proj"`` mode.
+    num_heads : int
+        Number of attention heads.
+    mlp_ratio : float, optional, default=4.0
+        Ratio of MLP hidden dimension to ``dim``.
+    qkv_bias : bool, optional, default=True
+        Whether attention QKV projections use a bias.
+    qk_norm : bool, optional, default=False
+        Whether to RMS-normalize attention queries and keys.
+    qk_norm_affine : bool, optional, default=False
+        Whether the QK RMS norms use a learnable affine scale.
+    mlp_drop_rate : float, optional, default=0.0
+        Dropout probability inside the MLP and attention output projection.
+    attn_drop_rate : float, optional, default=0.0
+        Dropout probability on attention weights.
+    attn_kernel : int | Tuple[int, int, int], optional, default=-1
+        Neighborhood-attention window; ``-1`` selects full attention.
+    na_dilation : int, optional, default=1
+        Dilation for neighborhood attention.
+    gated_attention : bool, optional, default=False
+        Whether attention outputs are multiplied by a learned sigmoid gate.
+    na3d_backend : str, optional, default=None
+        NATTEN backend forwarded to :class:`Natten3DSelfAttention`.
+    adaln_mode : Literal["pixel_proj", "bilinear_dw"], optional, default="pixel_proj"
+        How per-pixel modulation parameters are produced from ``backbone_cond``.
+    chunk_size_grouped_conv : int, optional, default=2
+        ``torch.vmap`` chunk size for the depthwise conv (``"bilinear_dw"`` only).
+    use_chunked_depthwise_conv : bool, optional, default=True
+        Whether the ``"bilinear_dw"`` depthwise conv uses the chunked
+        :class:`DepthwiseConv`; if ``False``, a plain grouped
+        :class:`torch.nn.Conv2d` is used.
+
+    Forward
+    -------
+    x : torch.Tensor
+        Pixel tokens of shape :math:`(B, D \cdot H \cdot W, \text{dim})`.
+    backbone_cond : torch.Tensor
+        Backbone conditioning tokens of shape :math:`(B, N_s, \text{cond\_dim})`.
+    pixel_dhw : Tuple[int, int, int]
+        Full pixel-grid shape :math:`(D, H, W)`.
+    backbone_dhw : Tuple[int, int, int]
+        Backbone patch-grid shape :math:`(s_d, s_h, s_w)`.
+    s_cond_bilinear : torch.Tensor, optional
+        Shared bilinear-upsampled conditioning (``"bilinear_dw"`` mode only),
+        as produced by :meth:`precompute_bilinear_cond`.
+    rope_tables : Tuple[torch.Tensor, torch.Tensor], optional
+        Precomputed ``(cos, sin)`` RoPE tables.
+
+    Outputs
+    -------
+    torch.Tensor
+        Pixel tokens of shape :math:`(B, D \cdot H \cdot W, \text{dim})`.
+    """
+
+    def __init__(
+        self,
+        dim: int,
+        cond_dim: int,
+        pixels_per_patch: int,
+        num_heads: int,
+        mlp_ratio: float = 4.0,
+        qkv_bias: bool = True,
+        qk_norm: bool = False,
+        qk_norm_affine: bool = False,
+        mlp_drop_rate: float = 0.0,
+        attn_drop_rate: float = 0.0,
+        attn_kernel: Union[int, Tuple[int, int, int]] = -1,
+        na_dilation: int = 1,
+        gated_attention: bool = False,
+        na3d_backend: Optional[str] = None,
+        adaln_mode: Literal["pixel_proj", "bilinear_dw"] = "pixel_proj",
+        chunk_size_grouped_conv: int = 2,
+        use_chunked_depthwise_conv: bool = True,
+    ):
+        super().__init__()
+        if adaln_mode not in ("pixel_proj", "bilinear_dw"):
+            raise ValueError(
+                f"adaln_mode must be 'pixel_proj' or 'bilinear_dw'; got {adaln_mode!r}"
+            )
+        self.adaln_mode = adaln_mode
+        self.num_adaln_params = 6  # shift1, scale1, gate1, shift2, scale2, gate2
+
+        self.attn = Natten3DSelfAttention(
+            dim,
+            num_heads=num_heads,
+            qkv_bias=qkv_bias,
+            qk_norm=qk_norm,
+            qk_norm_affine=qk_norm_affine,
+            attn_drop_rate=attn_drop_rate,
+            proj_drop_rate=mlp_drop_rate,
+            attn_kernel=attn_kernel,
+            do_depthwise_attention=False,
+            na_dilation=na_dilation,
+            gated_attention=gated_attention,
+            na3d_backend=na3d_backend,
+        )
+        self.norm1 = nn.LayerNorm(dim, elementwise_affine=False, eps=1e-6)
+        self.norm2 = nn.LayerNorm(dim, elementwise_affine=False, eps=1e-6)
+        self.mlp = Mlp(
+            in_features=dim,
+            hidden_features=int(dim * mlp_ratio),
+            out_features=dim,
+            act_layer=nn.GELU,
+            drop=mlp_drop_rate,
+        )
+
+        if adaln_mode == "bilinear_dw":
+            # Bilinear-upsample -> depthwise 5x5 smoothing -> RMSNorm -> GeLU ->
+            # zero-init projection, all at pixel resolution.
+            if use_chunked_depthwise_conv:
+                self.adaln_bilinear_dw_conv = DepthwiseConv(
+                    cond_dim,
+                    chunk_size=chunk_size_grouped_conv,
+                    kernel_size=(5, 5),
+                    padding=(2, 2),
+                    bias=True,
+                    padding_mode="replicate",
+                )
+            else:
+                self.adaln_bilinear_dw_conv = nn.Conv2d(
+                    cond_dim,
+                    cond_dim,
+                    kernel_size=(5, 5),
+                    padding=(2, 2),
+                    groups=cond_dim,
+                    bias=True,
+                    padding_mode="replicate",
+                )
+            # Identity-init: smoothing starts as a no-op (center tap = 1).
+            nn.init.zeros_(self.adaln_bilinear_dw_conv.weight)
+            self.adaln_bilinear_dw_conv.weight.data[:, 0, 2, 2] = 1.0
+            nn.init.zeros_(self.adaln_bilinear_dw_conv.bias)
+            self.adaln_bilinear_dw_norm = nn.RMSNorm(cond_dim, elementwise_affine=False)
+            self.adaln_bilinear_dw_proj = nn.Linear(
+                cond_dim, self.num_adaln_params * dim
+            )
+            nn.init.zeros_(self.adaln_bilinear_dw_proj.weight)
+            nn.init.zeros_(self.adaln_bilinear_dw_proj.bias)
+        else:
+            self.pixels_per_patch = pixels_per_patch
+            self.adaln_pixel_proj = nn.Sequential(
+                nn.SiLU(),
+                nn.Linear(cond_dim, pixels_per_patch * self.num_adaln_params * dim),
+            )
+            nn.init.zeros_(self.adaln_pixel_proj[-1].weight)
+            nn.init.zeros_(self.adaln_pixel_proj[-1].bias)
+
+    def reset_adaln_zero(self) -> None:
+        r"""Re-zero the AdaLN projection so the block starts as an identity residual.
+
+        The block zero-inits its AdaLN projection at construction; call this after
+        a blanket Xavier pass (e.g. in :meth:`Strata.initialize_weights`) that
+        would otherwise clobber it. The depthwise smoothing conv keeps its
+        identity init (a Conv2d, untouched by an ``nn.Linear`` Xavier pass).
+        """
+        if hasattr(self, "adaln_bilinear_dw_proj"):
+            nn.init.zeros_(self.adaln_bilinear_dw_proj.weight)
+            nn.init.zeros_(self.adaln_bilinear_dw_proj.bias)
+        if hasattr(self, "adaln_pixel_proj"):
+            nn.init.zeros_(self.adaln_pixel_proj[-1].weight)
+            nn.init.zeros_(self.adaln_pixel_proj[-1].bias)
+
+    @staticmethod
+    def _expand_cond_to_pixels(
+        adaln_raw: torch.Tensor,
+        pixel_dhw: Tuple[int, int, int],
+        backbone_dhw: Tuple[int, int, int],
+    ) -> torch.Tensor:
+        r"""Scatter per-patch modulation values to per-pixel order.
+
+        Parameters
+        ----------
+        adaln_raw : torch.Tensor
+            Per-patch modulation of shape
+            :math:`(B, N_s, p_{ppp} \cdot 6 \cdot \text{dim})`.
+        pixel_dhw : Tuple[int, int, int]
+            Full pixel-grid shape :math:`(D, H, W)`.
+        backbone_dhw : Tuple[int, int, int]
+            Backbone patch-grid shape :math:`(s_d, s_h, s_w)`.
+
+        Returns
+        -------
+        torch.Tensor
+            Per-pixel modulation of shape :math:`(B, D \cdot H \cdot W, 6 \cdot \text{dim})`.
+        """
+        d, h, w = pixel_dhw
+        sd, sh, sw = backbone_dhw
+        pv, ph, pw = d // sd, h // sh, w // sw
+        return rearrange(
+            adaln_raw,
+            "b (sd sh sw) (pv ph pw m) -> b (sd pv) (sh ph) (sw pw) m",
+            sd=sd,
+            sh=sh,
+            sw=sw,
+            pv=pv,
+            ph=ph,
+            pw=pw,
+        ).reshape(adaln_raw.shape[0], d * h * w, -1)
+
+    @staticmethod
+    def precompute_bilinear_cond(
+        backbone_cond: Float[torch.Tensor, "batch backbone_tokens cond_dim"],
+        pixel_dhw: Tuple[int, int, int],
+        backbone_dhw: Tuple[int, int, int],
+    ) -> Float[torch.Tensor, "batch depth cond_dim height width"]:
+        r"""Upsample backbone tokens to pixel resolution.
+
+        Shared across all ``"bilinear_dw"`` blocks so the (expensive) upsample is
+        computed once per forward pass. When the pixel depth equals the backbone
+        depth (:math:`s_d = D`, i.e. no depth upsampling) it uses a per-level **2D
+        bilinear** upsample — bit-identical to the original 2D-only path, so the
+        common case (backbone vertical patch size 1, or any matching pixel/backbone
+        depth) is numerically unchanged. This branch is deliberate, not just an
+        optimization: a 3D trilinear upsample at :math:`s_d = D` agrees with the
+        per-level 2D bilinear only to ~1e-7 (float), **not** exactly, so unifying on
+        trilinear would silently shift the numerics of every matching-depth model.
+        Only when the depth must actually be upsampled (:math:`s_d < D`) does it use
+        a **3D trilinear** upsample over :math:`(D, H, W)`.
+
+        Parameters
+        ----------
+        backbone_cond : torch.Tensor
+            Backbone tokens of shape :math:`(B, N_s, C)`.
+        pixel_dhw : Tuple[int, int, int]
+            Full pixel-grid shape :math:`(D, H, W)`.
+        backbone_dhw : Tuple[int, int, int]
+            Backbone patch-grid shape :math:`(s_d, s_h, s_w)`.
+
+        Returns
+        -------
+        torch.Tensor
+            Upsampled conditioning of shape :math:`(B, D, C, H, W)`.
+        """
+        d, h, w = pixel_dhw  # full pixel-grid depth / height / width
+        sd, sh, sw = backbone_dhw
+        if d == sd:
+            # Depth not upsampled (pixel depth == backbone depth): per-level 2D
+            # bilinear. Bit-identical to the original 2D path (backward compatible).
+            s_sp = rearrange(
+                backbone_cond, "b (sd sh sw) c -> (b sd) c sh sw", sd=sd, sh=sh, sw=sw
+            )
+            s_pix = F.interpolate(
+                s_sp, size=(h, w), mode="bilinear", align_corners=False
+            )
+            return rearrange(s_pix, "(b sd) c h w -> b sd c h w", sd=sd)
+        # Depth upsampled (pixel depth > backbone depth): 3D trilinear over (D,H,W).
+        s_sp = rearrange(backbone_cond, "b (sd sh sw) c -> b c sd sh sw", sd=sd, sh=sh, sw=sw)
+        s_pix = F.interpolate(
+            s_sp, size=(d, h, w), mode="trilinear", align_corners=False
+        )
+        return rearrange(s_pix, "b c d h w -> b d c h w")
+
+    def _compute_bilinear_dw_adaln_params(
+        self, s_cond_bilinear: torch.Tensor
+    ) -> Tuple[torch.Tensor, ...]:
+        r"""Compute the six per-pixel modulation tensors from upsampled conditioning.
+
+        Parameters
+        ----------
+        s_cond_bilinear : torch.Tensor
+            Trilinearly-upsampled conditioning of shape :math:`(B, D, C, H, W)`
+            (see :meth:`precompute_bilinear_cond`).
+
+        Returns
+        -------
+        Tuple[torch.Tensor, ...]
+            The six modulation tensors ``(shift1, scale1, gate1, shift2, scale2,
+            gate2)``, each of shape :math:`(B, D \cdot H \cdot W, \text{dim})`.
+        """
+        bsz, d, _, _, _ = s_cond_bilinear.shape
+        s_pix = rearrange(s_cond_bilinear, "b d c h w -> (b d) c h w")
+        s_pix = self.adaln_bilinear_dw_conv(s_pix)
+        # Channel-last RMSNorm, then back to channel-first.
+        s_pix = self.adaln_bilinear_dw_norm(s_pix.permute(0, 2, 3, 1)).permute(
+            0, 3, 1, 2
+        )
+        s_pix = rearrange(s_pix, "(b d) c h w -> b d c h w", b=bsz, d=d)
+        s_pix_seq = rearrange(s_pix, "b d c h w -> b (d h w) c")
+        s_pix_seq = F.gelu(s_pix_seq)
+        adaln_params = self.adaln_bilinear_dw_proj(s_pix_seq)
+        adaln_params = rearrange(
+            adaln_params, "b n (six c) -> b n six c", six=self.num_adaln_params
+        )
+        return adaln_params.unbind(dim=-2)
+
+    def forward(
+        self,
+        x: Float[torch.Tensor, "batch pixels dim"],
+        backbone_cond: Float[torch.Tensor, "batch patches cond_dim"],
+        pixel_dhw: Tuple[int, int, int],
+        backbone_dhw: Tuple[int, int, int],
+        s_cond_bilinear: Optional[torch.Tensor] = None,
+        rope_tables: Optional[RopeTables] = None,
+    ) -> Float[torch.Tensor, "batch pixels dim"]:
+        # Derive the six per-pixel modulation tensors from the conditioning.
+        if self.adaln_mode == "bilinear_dw":
+            if s_cond_bilinear is None:
+                raise ValueError(
+                    "s_cond_bilinear is required when adaln_mode='bilinear_dw'"
+                )
+            shift1, scale1, gate1, shift2, scale2, gate2 = (
+                self._compute_bilinear_dw_adaln_params(s_cond_bilinear)
+            )
+        else:
+            adaln_raw = self.adaln_pixel_proj(backbone_cond)  # (B, N_s, ppp*6*dim)
+            adaln_params = self._expand_cond_to_pixels(
+                adaln_raw, pixel_dhw, backbone_dhw
+            )  # (B, D*H*W, 6*dim)
+            shift1, scale1, gate1, shift2, scale2, gate2 = adaln_params.chunk(
+                self.num_adaln_params, dim=-1
+            )
+
+        # Attention sub-layer with modulated pre-norm and gated residual.
+        y = self.norm1(x) * (1 + scale1) + shift1
+        z = self.attn(y, latent_dhw=pixel_dhw, rope_tables=rope_tables)
+        x = x + gate1 * z
+
+        # MLP sub-layer with modulated pre-norm and gated residual.
+        y = self.norm2(x) * (1 + scale2) + shift2
+        z = self.mlp(y)
+        x = x + gate2 * z
+        return x
